Peer Recovery Support for DataFusion/Composite Engine (#20568)
* Peer recovery changes with softDelete policy

* Peer recovery MultiFileWriter changes

* Removed unnecessary changes and added test for peer recovery flow

* Removed unnecessary changes

* Revert changes to run.gradle and QueryPhase.java

* Peer recovery is working

* testDataFusionPrimaryRestartWithExtraCommits is failing

* Minor change in IndexShard for translogrecovery setting

* All tests related to parquet peer recovery and recovery is working

* Local peer recovery is working and all tests are passing

* Updated testPrimaryRelocationWithConcurrentIngestion for better check and assertion

* Refactoring PR

* Added fix for the scenario when parquet file written before compositeEngine creation with translog gets deleted

* nit fix

* updated the softDeletePolicy object creation parameter in compositeEngine

---------

+++ server/src/main/java/org/opensearch/index/engine/exec/coord/IndexFileDeleter.java
@@ -30,10 +30,16 @@ public class IndexFileDeleter {
     private final CompositeEngine compositeEngine;
 
     public IndexFileDeleter(CompositeEngine compositeEngine, CatalogSnapshot initialCatalogSnapshot, ShardPath shardPath) throws IOException {
+        this(compositeEngine, initialCatalogSnapshot, shardPath, true);
+    }
+
+    public IndexFileDeleter(CompositeEngine compositeEngine, CatalogSnapshot initialCatalogSnapshot, ShardPath shardPath, boolean deleteUnreferenced) throws IOException {
         this.compositeEngine = compositeEngine;
         if (initialCatalogSnapshot != null) {
             addFileReferences(initialCatalogSnapshot);
-            deleteUnreferencedFiles(shardPath);
+            if (deleteUnreferenced) {
+                deleteUnreferencedFiles(shardPath);
+            }
         }
     }
